Refactor Reservation and related DTOs for consistency and enhancements
- Change data types from Long to Integer for ID fields in entities and repositories.
- Refactor ActivityDto, CompanyDto, ReservationActivityDto, and ReservationDto for code consistency and clarity.
- Enhance ReservationDto to include `isCancelled` flag and replace `activities` with `reservedActivities`.
- Introduce `JsonInclude` annotations to exclude null fields from JSON responses.
- Extend Company entity from Guest and update fields to include first and last names.
- Add new methods in ReservationService and ReservationActivityService for handling reservation activities.
- Update repository interfaces to reflect changes in entity ID data types.
- Modify ReservationController to use Integer for ID path variables.

+++ src/main/java/dat3/adventureXP/api/ReservationController.java
@@ -31,19 +31,19 @@ public class ReservationController {
 
     // Read by ID (GET)
     @GetMapping("/{id}")
-    public ReservationDto getReservationById(@PathVariable Long id) {
+    public ReservationDto getReservationById(@PathVariable Integer id) {
         return reservationService.getReservationById(id);
     }
 
     // Update (PUT)
     @PutMapping("/{id}")
-    public ResponseEntity<ReservationDto> updateReservation(@PathVariable Long id, @RequestBody ReservationDto request) {
+    public ResponseEntity<ReservationDto> updateReservation(@PathVariable Integer id, @RequestBody ReservationDto request) {
         return reservationService.editReservation(id, request);
     }
 
     // Delete (DELETE)
     @DeleteMapping("/{id}")
-    public ResponseEntity<Reservation> deleteReservation(@PathVariable Long id) {
+    public ResponseEntity<Reservation> deleteReservation(@PathVariable Integer id) {
         return reservationService.deleteReservation(id);
     }
 }
