+++ src/main/java/dat3/adventureXP/repository/GuestRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface GuestRepository extends JpaRepository<Guest, Long> {
+public interface GuestRepository extends JpaRepository<Guest, Integer> {
     public Optional<Guest> findByFirstName(String firstName);
     public Optional<Guest> findByEmail(String email);
 }
